Split the BrowserToolbarController
